Inherit the `Node` namespace off a `NodePrototype`
Initializing a `Node` currently is a tricky process as it must exist in
the `ClusterAttr` for it to work correctly.

This creates a chicken and egg situation where a node object is required
in the `Configurator` before it exists. The `Configurator` uses this
object to find the higher level answers.

By splitting the objects up, it allows for greater flexibility in
resolving this issue
